refactor(app): lift the MQTT receiver and history loader out of AppContent
`AppContent` was a ~2000-line function holding 32 effects, and the app's
entire layout, in one body. Two self-contained pieces move out, verbatim.

`MqttLiveWiring` is the v2 streaming receiver: the connection lease, every
LiveEventEnvelope decode path, the turn-flush / interrupted-placeholder /
stale-stream machinery, and the session-live subscriptions. It turned out
to reach outside its own scope in exactly three places — `setMyActorId`
(now the `onMyActorId` prop), `currentTeamId` (now `teamId`), and one
`t()` call for a daemon error string (now its own `useTranslation`).

`SessionHistoryLoader` is local-first message-history loading. No props,
no return value, writes only to the session-message store.

Both render null and are mounted on *both* of `AppContent`'s return paths,
including the `!initialWorkspaceResolved` early return: as effects in the
component body these always ran, and a receiver that drops out while the
workspace resolves would be a real regression.

This is also a render fix. As a root-component subscription,
`useSessionListStore((s) => s.rows)` re-rendered the whole app on every
arriving message, because each one bumps `last_message_at` and re-sorts
the list into a fresh array. It now subscribes to a digest of the top
slice — the only thing that effect reacts to — and reads rows from
`getState()` when it runs. `mqttAccessToken`, the reconnect nonce, the
active session and the refresh triggers left the root along with it.

App.tsx 3,213 -> 1,339 lines; AppContent from 32 effects to 11. 100 now
unused imports pruned.

One behavioural difference worth knowing: React commits child effects
before parent effects, so this wiring now runs before `AppContent`'s
remaining effects rather than interleaved with them. Safe here because
every one of these effects is keyed on `userId`/`teamId` and bails while
they are null.
